* point plotting and waveforms all work -- going to reward myself by adding cluster-point-highlighting.
git-svn-id: file:///home/jonas/soma/svn-restore.20080710/software/tspikes/trunk@2154 5dccc5d9-beb7-0310-bf71-cc18d937a3f4

+++ src/glspikes.h
@@ -31,7 +31,7 @@ struct CRGBA_t
 struct GLSpikeWave_t
 {
   uint64_t ts; 
-  std::vector<int32_t> wave; 
+  std::vector<GLfloat> wave; 
 };
